+++ main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ProfileWrapper from './js/modules/profile.jsx';
+import ProfileWrapper from './js/modules/profileComponents/profile.jsx';
 
 ReactDOM.render(<ProfileWrapper />, document.getElementById('app'))
